Move Question and Report into Audits module
Also add a migration for updating the type column, as namespacing the
Question class has broken the single table inheritence.
