+++ app/models/audits/response.rb
@@ -8,8 +8,8 @@ module Audits
     validates :audit, presence: true
     validates :question, presence: true
 
-    scope :boolean, -> { joins(:question).where(questions: { type: "BooleanQuestion" }) }
-    scope :passing, -> { boolean.where(value: BooleanQuestion::PASS) }
+    scope :boolean, -> { joins(:question).where(questions: { type: "Audits::BooleanQuestion" }) }
+    scope :passing, -> { boolean.where(value: Audits::BooleanQuestion::PASS) }
     scope :failing, -> { boolean.where.not(id: passing) }
 
     validates :value, presence: { message: "mandatory field missing" },
